use online players count from proxy (b1.8 -> b1.7)

+++ src/main/java/com/github/dirtpowered/dirtmv/network/versions/Release73To61/metadata/V1_5RTo1_6RMetadataTransformer.java
@@ -71,7 +71,7 @@ public class V1_5RTo1_6RMetadataTransformer implements MetadataTransformer {
                 } else if (entityType == EntityType.HUMAN) {
                     if (index == 17) {
                         // absorption amount
-                        newMetaData.add(new WatchableObject(MetadataType.FLOAT, 17, ((Byte) value).floatValue()));
+                        newMetaData.add(new WatchableObject(MetadataType.FLOAT, 17, ((Integer) value).floatValue()));
                     }
 
                 } else {
